+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Fix GLIP Stat injection
 - Fix missing icon
+- Fix missing categories in `reportSunburstTicketByCategories` chart.
 
 ## [1.9.3] - 2026-01-08